[SYCL][NFC] Move exception_list class to separate file
