+++ sycl/include/CL/sycl/exception.hpp
@@ -13,8 +13,6 @@
 #include <CL/sycl/detail/common.hpp>
 #include <CL/sycl/stl.hpp>
 
-#include <cstddef>
-
 namespace cl {
 namespace sycl {
 
@@ -47,36 +45,6 @@ protected:
       : MMsg(Msg + " " + OCL_CODE_TO_STR(CLErr)), MCLErr(CLErr),
         MContext(Context) {}
 };
-
-// Forward declaration
-namespace detail {
-class queue_impl;
-}
-
-class exception_list {
-public:
-  using value_type = exception_ptr_class;
-  using reference = value_type &;
-  using const_reference = const value_type &;
-  using size_type = std::size_t;
-  using iterator = vector_class<exception_ptr_class>::const_iterator;
-  using const_iterator = vector_class<exception_ptr_class>::const_iterator;
-
-  size_type size() const;
-  // first asynchronous exception
-  iterator begin() const;
-  // refer to past-the-end last asynchronous exception
-  iterator end() const;
-
-private:
-  friend class detail::queue_impl;
-  void PushBack(const_reference Value);
-  void PushBack(value_type&& Value);
-  void Clear() noexcept;
-  vector_class<exception_ptr_class> MList;
-};
-
-using async_handler = function_class<void(cl::sycl::exception_list)>;
 
 class runtime_error : public exception {
 public:
