+++ sycl/source/detail/context_impl.cpp
@@ -12,6 +12,7 @@
 #include <CL/sycl/detail/clusm.hpp>
 #include <CL/sycl/device.hpp>
 #include <CL/sycl/exception.hpp>
+#include <CL/sycl/exception_list.hpp>
 #include <CL/sycl/info/info_desc.hpp>
 #include <CL/sycl/platform.hpp>
 #include <CL/sycl/stl.hpp>
