+++ sycl/source/exception.cpp
@@ -10,8 +10,6 @@
 #include <CL/sycl/context.hpp>
 #include <CL/sycl/exception.hpp>
 
-#include <utility>
-
 namespace cl {
 namespace sycl {
 
@@ -27,18 +25,6 @@ context exception::get_context() const {
 }
 
 cl_int exception::get_cl_code() const { return MCLErr; }
-
-exception_list::size_type exception_list::size() const { return MList.size(); }
-
-exception_list::iterator exception_list::begin() const { return MList.begin(); }
-
-exception_list::iterator exception_list::end() const { return MList.cend(); }
-
-void exception_list::PushBack(const_reference Value) { MList.emplace_back(Value); }
-
-void exception_list::PushBack(value_type&& Value) { MList.emplace_back(std::move(Value)); }
-
-void exception_list::Clear() noexcept { MList.clear(); }
 
 } // namespace sycl
 } // namespace cl
